Add cast ending move speed to int

+++ Managers/ObjectSpawner.cs
@@ -264,7 +264,7 @@ public class ObjectSpawner : MonoBehaviour {
 
 	public void SetEndingMoveSpeedForHighScoreEntry() {
 		float endingMoveSpeedKPH = currentObjectSpeed * 3.6f;
-		cachedObjectSpeed = endingMoveSpeedKPH / 1.609f;
+		cachedObjectSpeed = (int)endingMoveSpeedKPH / 1.609f;
 	}
 
 	public void SetEndingSpawnSpeedForHighScoreEntry() {
